+++ joulescope_ui/widgets/multimeter/meter_widget.py
@@ -12,12 +12,36 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from PySide2 import QtCore, QtWidgets
+from PySide2 import QtCore, QtGui, QtWidgets
 from joulescope_ui import joulescope_rc
+from joulescope_ui.template import render
 from joulescope.units import three_sig_figs
 from .meter_value_widget import MeterValueWidget
+from joulescope_ui.ui_util import rgba_to_css
 import logging
 log = logging.getLogger(__name__)
+
+
+STYLESHEET = """\
+QWidget[multimeter=true] {
+  padding: 0px;
+  color: {% multimeter_color %};
+}
+
+QWidget[multimeter_spacer=true] {
+  background-color: {% multimeter_background_half %};
+}
+
+QLabel[multimeter_label=true] {
+  padding-left: 2px;
+  padding-right: 2px;
+  padding-top: 2px;
+  padding-bottom: 2px; 
+  background-color: {% multimeter_background_color %}; 
+  color: {% multimeter_color %};
+  border-color: white;
+}
+"""
 
 
 FIELDS = [
@@ -34,6 +58,7 @@ class MeterWidget(QtWidgets.QWidget):
         QtWidgets.QWidget.__init__(self, parent)
         self._cmdp = cmdp
         self._accumulate_duration = 0.0
+
         self.verticalLayout = QtWidgets.QVBoxLayout(self)
         self.verticalLayout.setObjectName("verticalLayout")
         self.verticalLayout.setSpacing(0)
@@ -54,22 +79,78 @@ class MeterWidget(QtWidgets.QWidget):
         self.controlSpacer = QtWidgets.QSpacerItem(40, 20, QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Minimum)
         self.controlLayout.addItem(self.controlSpacer)
 
+        self._grid_widget = QtWidgets.QWidget(self)
+        self.verticalLayout.addWidget(self._grid_widget)
+        self._grid_widget.setProperty('multimeter', True)
+
+        self._grid_layout = QtWidgets.QGridLayout(self._grid_widget)
+        self._grid_layout.setContentsMargins(0, 0, 0, 0)
+        self._grid_layout.setMargin(0)
+        self._grid_layout.setHorizontalSpacing(0)
+        self._grid_layout.setVerticalSpacing(0)
+
+        self.spacers = []
         self.values = {}
-        for name, units_short, units_long in FIELDS:
-            w = MeterValueWidget(self, cmdp)
-            w.setStyleSheet("QWidget { background-color : black; color : green; }")
+        for idx, (name, units_short, units_long) in enumerate(FIELDS):
+            if idx:
+                spacer = QtWidgets.QWidget(self._grid_widget)
+                spacer.setProperty('multimeter_spacer', True)
+                spacer.setSizePolicy(QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Minimum)
+                spacer.setMinimumHeight(2)
+                self._grid_layout.addWidget(spacer, idx * 5 - 1, 0, 1, 4)
+                self.spacers.append(spacer)
+            w = MeterValueWidget(self._grid_widget, cmdp, idx * 5, name)
             w.configure(name.capitalize(), units_short, units_long)
             self.values[name] = w
-            w.setContentsMargins(0, 0, 0, 0)
-            self.verticalLayout.addWidget(w)
         self.values['energy'].configure_energy()
 
+        self._grid_layout.setColumnStretch(0, 1)
+        self._grid_layout.setColumnStretch(1, 1)
         self.sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Expanding)
         self.sizePolicy.setHorizontalStretch(0)
         self.sizePolicy.setVerticalStretch(0)
         self.setSizePolicy(self.sizePolicy)
         self.retranslateUi()
         self._cmdp.subscribe('Device/#state/statistics', self._on_device_statistics, update_now=True)
+
+        cmdp.subscribe('Widgets/Multimeter/font-main', self._on_font_main, update_now=True)
+        cmdp.subscribe('Widgets/Multimeter/font-stats', self._on_font_stats, update_now=True)
+        cmdp.subscribe('Widgets/Multimeter/font-color', self._on_color)
+        cmdp.subscribe('Widgets/Multimeter/background-color', self._on_color, update_now=True)
+
+    def _on_font_main(self, topic, value):
+        font = QtGui.QFont()
+        font.fromString(value)
+        metrics = QtGui.QFontMetrics(font)
+        w1 = metrics.boundingRect("i+0.00000i").width()
+        w2 = metrics.boundingRect("iMW").width()
+        for value in self.values.values():
+            for w, width in zip(value.main_widgets, [w1, w2]):
+                w.setFont(font)
+                w.setMinimumWidth(width)
+
+    def _on_font_stats(self, topic, value):
+        font = QtGui.QFont()
+        font.fromString(value)
+        metrics = QtGui.QFontMetrics(font)
+        width = metrics.boundingRect("i+0.00000i").width()
+        for value in self.values.values():
+            for w1, w2 in value.stats_widgets:
+                w1.setFont(font)
+                w1.setMinimumWidth(width)
+                w2.setFont(font)
+
+    def _on_color(self, topic, value):
+        foreground = rgba_to_css(self._cmdp['Widgets/Multimeter/font-color'])
+        bg = self._cmdp['Widgets/Multimeter/background-color']
+        background = rgba_to_css(bg)
+        r, g, b, a = bg
+        bg_half = rgba_to_css([r, g, b, a * 0.75])
+        style = render(STYLESHEET,
+                       multimeter_background_color=background,
+                       multimeter_color=foreground,
+                       multimeter_background_half=bg_half)
+        self._grid_widget.setStyleSheet(style)
 
     @QtCore.Slot(bool)
     def on_accumulate_toggled(self, checked):