+++ CHANGELOG.md
@@ -35,6 +35,7 @@ This file contains the list of changes made to pyjoulescope_ui.
 *   Added on/off switch #84.
 *   Added horizontal markers to the waveform widget #37.
 *   Added statistics on/off for all vertical markers #12.
+*   Improved multimeter view grid layout.
 
 
 ## 0.8.16